Get rid of renameTo atomic copy fallback
There should be no case where renameTo fails, but you want to just go ahead with the atomic copy.  If it fails it is because something still has a lock on the file, and in that case the atomicCopy fails as well and throws the IOError up and causes other issues.