Fix for recycling old transition drawable
The old implementation only kept current + 1 old drawable.  While this reduced the number of cases where the old drawable was attempted to be used, it did not completely eliminate it.  There were still cases where the + 1 old drawable was used and caused a crash.  This new implementation just completely clears the old transition drawable from the stack after a certain amount of time.