If nextSetup is true, use playNext regardless
If nextSetup is true, then the media player has already called setNextMediaPlayer.  Since this is already inside of the onCompletion, there is a pretty good chance that the next media player is already playing.  In this case we can't risk going into the condition where the current media player will be restarted.