Rework temp file logic to fix erroneous cache reset, add more indepth logging
createNewFile actually returns false if file already exists, not if the file failed to be created.  If it fails, it should be throwing an exception.  It already existing is not a good enough basis to fail and to cause the user issues.  Also changed it to fail if it is not able to re-delete the file.