Fix issues with playlist file listings in offline mode
In offline mode, playlists are read from a .m3u file which stores the absolute paths.  No matter what, the files need to be passed to createEntry without the .complete part in the filename or else the files will not be found.  If the normal name file does not exist, need to check against the .complete one as well before removing it from the list of valid songs to show in the playlist.