#415 Improve bookmark logic when playing a series of audio books
For audio books we want to modify the automatic creation of deletion of bookmarks.  An example of what this is supposed to handle:

Start in chapter 8 from bookmark, continue to chapter to 9 of 10.  At position 8:50 of 10:00.  Start playing some music.  Current behavior is that this will have deleted a bookmark from 8 and not create one at 9.  Now, it will create one at 9 since there is still another chapter to read.