Remove registerMediaButtonEventReceiver for API >= 21
Finally figured out that this was causing an extra controller to be
registered with the AVRCP module. With DSub as the only media app
running the log would show `V/NewAvrcpMediaPlayerList:
onActiveSessionsChanged: number of controllers: 0` on entry into the app
and would increase to 2 shortly after.

On Android Pie this causes problems as the new AVRCP module's logic is
causing the metadata to be sent through to the wrong controller. I'm not
sure if you think I've gone about it in the correct way but I've removed
the setting for media buttons as well for Lollipop onwards, considering
that this setting has no effect anyway on >= 21.