Settings: Remove overly zealous URL validation
Server URLs were checked if they contain an '@' or ' ' (space) and
rejected. This is wrong, an URL may contain an @ sign to provide
basic-auth parameters as part of the URL and works just fine.

We should leave URL validation to libraries dealing with URLs. With this
change it's now possible to connect to a subsonic server protected by
basic-auth, as long as the basic-auth credentials match the subsonic
credentials.