Remove @ validation from network settings
Was a PR to completely remove space and @ validation (https://github.com/daneren2005/Subsonic/pull/985), but I can see people easily pressing space without meaning to and not being able to see what is wrong.  If someone makes a good reason for it to be needed I will remove it, but for now just removing @ so people can do basic auth.